Add inquiry page
